Finished
Finished my sudoku solver API application. I added an animated background, changed the color scheme, cleaned some of the layout, inputs, and code up. I don't think I'm going to convert this to a react app, not anytime soon anyway. I'm happy with it the way it is right now. Next I'm going to look in to deploying it and making a live demo so that I can showcase it on my web portfolio.